mux/mux.go
package mux

import (
	"net/http"
	"strings"
)

// Subdomains will store all of registered subdomains
type Subdomains map[string]http.Handler

func (subdomains Subdomains) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	domainParts := strings.Split(r.Host, ".")

	if mux := subdomains[domainParts[0]]; mux != nil {
		// Let the appropriate mux serve the request
		mux.ServeHTTP(w, r)
	} else {
		// Handle 404
		http.Error(w, "Not found", 404)
	}
}

// Mux is request multiplexer
type Mux struct {
	http.Handler
}
